test: introduce tests/common and move go interop there
